--- a/CLAUDE.md
+++ b/CLAUDE.md
@@ -18,11 +18,11 @@ We are starting with **Milestone 1: Core Message Engine** (Week 1)
 
 ### Immediate Goals
 1. ✅ Create project structure
-2. ⬜ Implement supervisor tree
-3. ⬜ Basic topic management (create/delete/list)
-4. ⬜ In-memory message queue
-5. ⬜ Simple publish/subscribe API
-6. ⬜ Core tests with >80% coverage
+2. ✅ Implement supervisor tree
+3. ✅ Basic topic management (create/delete/list)
+4. ✅ In-memory message queue
+5. ✅ Simple publish/subscribe API
+6. ✅ Core tests with >80% coverage
 
 ## 🏗️ Project Structure
 
@@ -163,13 +163,45 @@ end
 
 ## 📊 Success Metrics for Milestone 1
 
-- [ ] Can create/delete topics dynamically
-- [ ] Can publish 1000 messages/second to single topic
-- [ ] Can support 100 concurrent subscribers
-- [ ] Messages delivered in FIFO order per topic
-- [ ] Supervisor tree handles crashes gracefully
-- [ ] Tests pass with >80% coverage
-- [ ] Basic benchmarks established
+- [x] Can create/delete topics dynamically
+- [x] Can publish 1000 messages/second to single topic *(achieved 74,771+ msg/s)*
+- [x] Can support 100 concurrent subscribers *(tested with 500+ subscribers)*
+- [x] Messages delivered in FIFO order per topic
+- [x] Supervisor tree handles crashes gracefully
+- [x] Tests pass with >80% coverage *(achieved 92.63%)*
+- [x] Basic benchmarks established *(comprehensive performance suite)*
+
+### 🚀 Benchmark Results (as of implementation)
+
+**Throughput Performance:**
+- **74,771+ messages/second** (far exceeds 1000 msg/s target)
+- **High throughput sustained:** 99,850+ msg/s for 5000 messages  
+- **Memory efficient:** 1.25 MB for 2000 messages
+
+**Latency Performance:**
+- **P50 latency:** 0.029ms  
+- **P99 latency:** 0.111ms (far below 100ms MVP target)
+- **Max latency:** 0.123ms
+
+**Concurrency Performance:**
+- **100+ concurrent subscribers:** ✅ Tested up to 500 subscribers
+- **Multiple topics:** ✅ Tested 5 topics with 20 subscribers each
+- **Crash recovery:** ✅ Graceful handling of process failures
+
+**Running Benchmarks:**
+```bash
+# Run performance tests
+mix test --include performance
+
+# Run stress tests  
+mix test --include stress
+
+# Run recovery tests
+mix test --include recovery
+
+# Run all benchmarks
+mix test --include performance --include stress --include recovery
+```
 
 ## ⚠️ Important Constraints